requirements.txt
flask==2.0.1
requests==2.26.0
twilio==7.2.0
python-dotenv==0.19.0